fix: improve full page forms to avoid using a modal
