+++ assets/svelte/consumers/SinkConsumerForm.svelte
@@ -16,7 +16,7 @@
     ExpandableCard,
   } from "$lib/components/ui/card";
   import { Label } from "$lib/components/ui/label";
-  import FullPageModal from "../components/FullPageModal.svelte";
+  import FullPageForm from "../components/FullPageForm.svelte";
   import { cn } from "$lib/utils";
   import FilterForm from "../components/FilterForm.svelte";
   import GroupColumnsForm from "./GroupColumnsForm.svelte";
@@ -37,6 +37,7 @@
   import ElasticsearchSinkForm from "$lib/sinks/elasticsearch/ElasticsearchSinkForm.svelte";
   import * as Alert from "$lib/components/ui/alert/index.js";
   import TableSelector from "../components/TableSelector.svelte";
+  import * as Tooltip from "$lib/components/ui/tooltip";
   import * as Popover from "$lib/components/ui/popover";
   import * as Dialog from "$lib/components/ui/dialog";
   import MessageExamples from "$lib/components/MessageExamples.svelte";
@@ -253,8 +254,6 @@
   }
 
   const isEditMode = !!consumer.id;
-  let dialogOpen = true;
-  let showConfirmDialog = false;
 
   function handleConsumerSubmit() {
     isSubmitting = true;
@@ -382,16 +381,14 @@
   const exampleUnixMicrosecondTimestamp = 1704086400000000;
 </script>
 
-<FullPageModal
+<FullPageForm
   title={isEditMode ? `Edit ${consumerTitle}` : `Create ${consumerTitle}`}
-  bind:open={dialogOpen}
-  bind:showConfirmDialog
   showConfirmOnExit={isDirty}
   on:close={handleClose}
 >
   <form
     on:submit|preventDefault={handleConsumerSubmit}
-    class="space-y-6 max-w-3xl mx-auto mt-6"
+    class="space-y-6 max-w-3xl mx-auto"
   >
     <Card>
       <CardHeader>
@@ -527,29 +524,17 @@
     <ExpandableCard disabled={!transformSectionEnabled} expanded={!isEditMode}>
       <svelte:fragment slot="title">
         Transforms
-        <button on:click|stopPropagation>
-          <Popover.Root>
-            <Popover.Trigger asChild let:builder>
-              <Button
-                builders={[builder]}
-                variant="link"
-                class="text-muted-foreground hover:text-foreground p-0"
-              >
-                <Info class="h-4 w-4" />
-              </Button>
-            </Popover.Trigger>
-            <Popover.Content class="w-80">
-              <div class="grid gap-4">
-                <div class="space-y-2">
-                  <p class="text-sm text-muted-foreground font-normal">
-                    Transform your messages before they are sent to the sink
-                    destination.
-                  </p>
-                </div>
-              </div>
-            </Popover.Content>
-          </Popover.Root>
-        </button>
+        <Tooltip.Root openDelay={200}>
+          <Tooltip.Trigger>
+            <Info class="h-4 w-4 text-gray-400 cursor-help" />
+          </Tooltip.Trigger>
+          <Tooltip.Content class="p-4 max-w-xs">
+            <p class="text-sm text-muted-foreground font-normal">
+              Transform your messages before they are sent to the sink
+              destination.
+            </p>
+          </Tooltip.Content>
+        </Tooltip.Root>
         <Beta size="sm" variant="subtle" />
       </svelte:fragment>
 
@@ -601,35 +586,23 @@
         <svelte:fragment slot="title">
           Initial backfill
 
-          <button on:click|stopPropagation>
-            <Popover.Root>
-              <Popover.Trigger asChild let:builder>
-                <Button
-                  builders={[builder]}
-                  variant="link"
-                  class="text-muted-foreground hover:text-foreground p-0"
+          <Tooltip.Root openDelay={200}>
+            <Tooltip.Trigger>
+              <Info class="h-4 w-4 text-gray-400 cursor-help" />
+            </Tooltip.Trigger>
+            <Tooltip.Content class="p-4 max-w-xs">
+              <p class="text-sm text-muted-foreground font-normal">
+                Sequin will run an initial <a
+                  href="https://sequinstream.com/docs/reference/backfills"
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  class="text-primary underline"
                 >
-                  <Info class="h-4 w-4" />
-                </Button>
-              </Popover.Trigger>
-              <Popover.Content class="w-80">
-                <div class="grid gap-4">
-                  <div class="space-y-2">
-                    <p class="text-sm text-muted-foreground font-normal">
-                      Sequin will run an initial <a
-                        href="https://sequinstream.com/docs/reference/backfills"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        class="text-primary underline"
-                      >
-                        backfill
-                      </a> of data from the selected table to the sink destination.
-                    </p>
-                  </div>
-                </div>
-              </Popover.Content>
-            </Popover.Root>
-          </button>
+                  backfill
+                </a> of data from the selected table to the sink destination.
+              </p>
+            </Tooltip.Content>
+          </Tooltip.Root>
         </svelte:fragment>
 
         <svelte:fragment slot="summary">
@@ -823,36 +796,38 @@
         <div class="flex justify-end items-center gap-2">
           {#if consumer.type !== "http_push" && consumer.type !== "sequin_stream"}
             <Popover.Root bind:open={showLocalhostWarningDialog}>
-              <Popover.Trigger />
-              <Button
-                loading={testConnectionState.status === "loading"}
-                variant="outline"
-                class="self-end"
-                on:click={onTestConnection}
-              >
-                {#if testConnectionState.status === "success" && testConnectionState.displayStatus}
-                  <span
-                    class="flex items-center p-1 gap-1 mr-2 bg-green-500 rounded-full"
-                  ></span>
-                  Connection succeeded
-                {:else if testConnectionState.status === "error" && testConnectionState.displayStatus}
-                  <span
-                    class="flex items-center p-1 gap-1 mr-2 bg-red-500 rounded-full"
-                  ></span>
-                  Connection failed
-                {:else}
-                  <span
-                    class="flex items-center w-2 h-2 mr-2 rounded-full"
-                    class:bg-green-500={testConnectionState.lastTestStatus ===
-                      "success"}
-                    class:bg-red-500={testConnectionState.lastTestStatus ===
-                      "error"}
-                    class:bg-gray-300={testConnectionState.lastTestStatus ===
-                      "none"}
-                  ></span>
-                  Test Connection
-                {/if}
-              </Button>
+              <Popover.Trigger asChild let:builder>
+                <Button
+                  loading={testConnectionState.status === "loading"}
+                  variant="outline"
+                  class="self-end"
+                  builders={[builder]}
+                  on:click={onTestConnection}
+                >
+                  {#if testConnectionState.status === "success" && testConnectionState.displayStatus}
+                    <span
+                      class="flex items-center p-1 gap-1 mr-2 bg-green-500 rounded-full"
+                    ></span>
+                    Connection succeeded
+                  {:else if testConnectionState.status === "error" && testConnectionState.displayStatus}
+                    <span
+                      class="flex items-center p-1 gap-1 mr-2 bg-red-500 rounded-full"
+                    ></span>
+                    Connection failed
+                  {:else}
+                    <span
+                      class="flex items-center w-2 h-2 mr-2 rounded-full"
+                      class:bg-green-500={testConnectionState.lastTestStatus ===
+                        "success"}
+                      class:bg-red-500={testConnectionState.lastTestStatus ===
+                        "error"}
+                      class:bg-gray-300={testConnectionState.lastTestStatus ===
+                        "none"}
+                    ></span>
+                    Test Connection
+                  {/if}
+                </Button>
+              </Popover.Trigger>
               <Popover.Content class="w-80">
                 <div class="grid gap-4">
                   <div class="space-y-2">
@@ -928,7 +903,7 @@
       </CardContent>
     </Card>
   </form>
-</FullPageModal>
+</FullPageForm>
 
 <Dialog.Root bind:open={showMessageTypeExampleModal}>
   <Dialog.Portal>
