+++ lib/sequin_web/live/sink_consumers/index.ex
@@ -3,25 +3,10 @@ defmodule SequinWeb.SinkConsumersLive.Index do
   use SequinWeb, :live_view
 
   alias Sequin.Consumers
-  alias Sequin.Consumers.AzureEventHubSink
-  alias Sequin.Consumers.ElasticsearchSink
-  alias Sequin.Consumers.GcpPubsubSink
-  alias Sequin.Consumers.HttpPushSink
-  alias Sequin.Consumers.KafkaSink
-  alias Sequin.Consumers.NatsSink
-  alias Sequin.Consumers.RabbitMqSink
-  alias Sequin.Consumers.RedisStreamSink
-  alias Sequin.Consumers.RedisStringSink
-  alias Sequin.Consumers.SequinStreamSink
   alias Sequin.Consumers.SinkConsumer
-  alias Sequin.Consumers.SnsSink
-  alias Sequin.Consumers.SqsSink
-  alias Sequin.Consumers.TypesenseSink
   alias Sequin.Databases
-  alias Sequin.Databases.DatabaseUpdateWorker
   alias Sequin.Health
   alias Sequin.Metrics
-  alias SequinWeb.Components.ConsumerForm
   alias SequinWeb.RouteHelpers
 
   @smoothing_window 5
@@ -68,17 +53,8 @@ defmodule SequinWeb.SinkConsumersLive.Index do
   end
 
   @impl Phoenix.LiveView
-  def render(%{live_action: :new} = assigns) do
-    ~H"""
-    <div id="consumers-index">
-      <%= render_consumer_form(assigns) %>
-    </div>
-    """
-  end
-
   def render(assigns) do
     consumers = Enum.filter(assigns.consumers, &is_struct(&1, SinkConsumer))
-
     encoded_consumers = Enum.map(consumers, &encode_consumer/1)
 
     assigns = assign(assigns, :encoded_consumers, encoded_consumers)
@@ -111,190 +87,10 @@ defmodule SequinWeb.SinkConsumersLive.Index do
     |> assign(:live_action, :index)
   end
 
-  defp apply_action(socket, :new, %{"kind" => kind}) do
-    # Refresh tables for all databases in the account
-    account = current_account(socket)
-    databases = Databases.list_dbs_for_account(account.id)
-    Enum.each(databases, &DatabaseUpdateWorker.enqueue(&1.id))
-
-    socket
-    |> assign(:page_title, "New Sink")
-    |> assign(:live_action, :new)
-    |> assign(:form_kind, kind)
-  end
-
-  defp render_consumer_form(%{form_kind: "http_push"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :http_push, sink: %HttpPushSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "sqs"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :sqs, sink: %SqsSink{}, batch_size: 10}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "sns"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :sns, sink: %SnsSink{}, batch_size: 10}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "kafka"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :kafka, batch_size: 10, sink: %KafkaSink{tls: false}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "redis_stream"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :redis_stream, batch_size: 10, sink: %RedisStreamSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "sequin_stream"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :sequin_stream, sink: %SequinStreamSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "gcp_pubsub"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={
-        %SinkConsumer{
-          type: :gcp_pubsub,
-          sink: %GcpPubsubSink{}
-        }
-      }
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "nats"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :nats, sink: %NatsSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "rabbitmq"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :rabbitmq, sink: %RabbitMqSink{virtual_host: "/"}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "azure_event_hub"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :azure_event_hub, sink: %AzureEventHubSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "typesense"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :typesense, sink: %TypesenseSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "elasticsearch"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :elasticsearch, sink: %ElasticsearchSink{}}}
-    />
-    """
-  end
-
-  defp render_consumer_form(%{form_kind: "redis_string"} = assigns) do
-    ~H"""
-    <.live_component
-      current_user={@current_user}
-      module={ConsumerForm}
-      id="new-consumer"
-      action={:new}
-      consumer={%SinkConsumer{type: :redis_string, batch_size: 10, sink: %RedisStringSink{}}}
-    />
-    """
-  end
-
   @impl Phoenix.LiveView
   def handle_info(:update_consumers, socket) do
     Process.send_after(self(), :update_consumers, 1000)
     {:noreply, assign(socket, :consumers, load_consumers(socket))}
-  end
-
-  def handle_info({:database_tables_updated, _updated_database}, socket) do
-    # Proxy down to ConsumerForm
-    send_update(ConsumerForm, id: "new-consumer", event: :database_tables_updated)
-
-    {:noreply, socket}
   end
 
   defp load_consumers(socket) do
