+++ lib/sequin_web/router.ex
@@ -93,7 +93,7 @@ defmodule SequinWeb.Router do
 
     live_session :default, on_mount: [{SequinWeb.UserAuth, :ensure_authenticated}, {SequinWeb.LiveHooks, :global}] do
       live "/sinks", SinkConsumersLive.Index, :index
-      live "/sinks/new", SinkConsumersLive.Index, :new
+      live "/sinks/new", SinkConsumersLive.Form, :new
       live "/sinks/:type/:id", SinkConsumersLive.Show, :show
       live "/sinks/:type/:id/messages", SinkConsumersLive.Show, :messages
       live "/sinks/:type/:id/edit", SinkConsumersLive.Show, :edit
